Add Wrath version of CMSG_MOVE_TIME_SKIPPED with PackedGuid
