+++ wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped_3_3_5.wowm
@@ -1,0 +1,6 @@
+cmsg CMSG_MOVE_TIME_SKIPPED = 0x02CE {
+    PackedGuid guid;
+    u32 lag;
+} {
+    versions = "3.3.5";
+}
