+++ wow_world_messages/src/world/shared/mod.rs
@@ -112,7 +112,7 @@ pub mod cmsg_mail_take_money_vanilla_tbc_wrath;
 pub mod cmsg_meetingstone_info_vanilla_tbc;
 pub mod cmsg_mountspecial_anim_vanilla_tbc_wrath;
 pub mod cmsg_move_set_raw_position_vanilla_tbc_wrath;
-pub mod cmsg_move_time_skipped_vanilla_tbc_wrath;
+pub mod cmsg_move_time_skipped_vanilla_tbc;
 pub mod cmsg_name_query_vanilla_tbc_wrath;
 pub mod cmsg_next_cinematic_camera_vanilla_tbc_wrath;
 pub mod cmsg_offer_petition_tbc_wrath;
