+++ wow_world_messages/src/world/wrath/cmsg_move_time_skipped.rs
@@ -1,0 +1,59 @@
+use std::convert::{TryFrom, TryInto};
+use crate::Guid;
+use std::io::{Write, Read};
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
+/// Auto generated from the original `wowm` in file [`wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped_3_3_5.wowm:1`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped_3_3_5.wowm#L1):
+/// ```text
+/// cmsg CMSG_MOVE_TIME_SKIPPED = 0x02CE {
+///     PackedGuid guid;
+///     u32 lag;
+/// }
+/// ```
+pub struct CMSG_MOVE_TIME_SKIPPED {
+    pub guid: Guid,
+    pub lag: u32,
+}
+
+impl crate::Message for CMSG_MOVE_TIME_SKIPPED {
+    const OPCODE: u32 = 0x02ce;
+
+    fn size_without_header(&self) -> u32 {
+        self.size() as u32
+    }
+
+    fn write_into_vec(&self, w: &mut Vec<u8>) -> Result<(), std::io::Error> {
+        let size_assert_header_size = w.len();
+        // guid: PackedGuid
+        self.guid.write_packed_guid_into_vec(w);
+
+        // lag: u32
+        w.write_all(&self.lag.to_le_bytes())?;
+
+        assert_eq!(self.size() as usize + size_assert_header_size, w.len(), "Mismatch in pre-calculated size and actual written size. This needs investigation as it will cause problems in the game client when sent");
+        Ok(())
+    }
+    fn read_body(r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
+        // guid: PackedGuid
+        let guid = Guid::read_packed(r)?;
+
+        // lag: u32
+        let lag = crate::util::read_u32_le(r)?;
+
+        Ok(Self {
+            guid,
+            lag,
+        })
+    }
+
+}
+#[cfg(feature = "wrath")]
+impl crate::world::wrath::ClientMessage for CMSG_MOVE_TIME_SKIPPED {}
+
+impl CMSG_MOVE_TIME_SKIPPED {
+    pub(crate) fn size(&self) -> usize {
+        self.guid.size() // guid: Guid
+        + 4 // lag: u32
+    }
+}
+
