+++ wowm_language/src/docs/cmsg_move_time_skipped.md
@@ -1,10 +1,10 @@
 # CMSG_MOVE_TIME_SKIPPED
 
-## Client Version 1, Client Version 2, Client Version 3
+## Client Version 1, Client Version 2
 
 ### Wowm Representation
 
-Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped.wowm:3`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped.wowm#L3).
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped.wowm:1`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped.wowm#L1).
 ```rust,ignore
 cmsg CMSG_MOVE_TIME_SKIPPED = 0x02CE {
     Guid guid;
@@ -28,4 +28,35 @@ CMSG have a header of 6 bytes.
 | ------ | ----------------- | ---- | ---- | ----------- | ------- |
 | 0x06 | 8 / Little | [Guid](../spec/packed-guid.md) | guid |  |  |
 | 0x0E | 4 / Little | u32 | lag |  |  |
+
+# CMSG_MOVE_TIME_SKIPPED
+
+## Client Version 3.3.5
+
+### Wowm Representation
+
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped_3_3_5.wowm:1`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/cmsg/cmsg_move_time_skipped_3_3_5.wowm#L1).
+```rust,ignore
+cmsg CMSG_MOVE_TIME_SKIPPED = 0x02CE {
+    PackedGuid guid;
+    u32 lag;
+}
+```
+### Header
+
+CMSG have a header of 6 bytes.
+
+#### CMSG Header
+
+| Offset | Size / Endianness | Type   | Name   | Description |
+| ------ | ----------------- | ------ | ------ | ----------- |
+| 0x00   | 2 / Big           | uint16 | size   | Size of the rest of the message including the opcode field but not including the size field.|
+| 0x02   | 4 / Little        | uint32 | opcode | Opcode that determines which fields the message contains.|
+
+### Body
+
+| Offset | Size / Endianness | Type | Name | Description | Comment |
+| ------ | ----------------- | ---- | ---- | ----------- | ------- |
+| 0x06 | - / - | [PackedGuid](../spec/packed-guid.md) | guid |  |  |
+| - | 4 / Little | u32 | lag |  |  |
